city search, retrieving current weather

+++ app/src/main/java/gmoby/android/weatherapp/Model/Weather/Clouds.java
@@ -10,13 +10,13 @@ import com.google.gson.annotations.SerializedName;
 public class Clouds {
     @SerializedName("all")
     @Expose
-    private Integer all;
+    private Double all;
 
-    public Integer getAll() {
+    public Double getAll() {
         return all;
     }
 
-    public void setAll(Integer all) {
+    public void setAll(Double all) {
         this.all = all;
     }
 }
